% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.combn.R
\name{count.combn}
\alias{count.combn}
\title{Count paired comparisons}
\usage{
count.combn(nitem)
}
\arguments{
\item{nitem}{Number of items.}
}
\value{
An integer corresponding to the number of paired comparisons.
}
\description{
This function calculates the number of paired comparisons needed to compare a
 set of N items.
}
\details{
This function is only useful, if the number of paired comparisons of a full
design, that is all possible paired comparisons, is of interest. Then the 
number is
\deqn{\frac{N\times(N-1)}{2}}
}
\examples{

# Number if paired comparisons for a set of 15 items = 105.
count.combn(15)

}
