% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_FC_raw.R
\docType{data}
\name{FC_raw}
\alias{FC_raw}
\title{Raw ranking data of $N=15$ items from three factors/traits (Thurstonian modeling)}
\format{
A data frame with 1000 observations on 15 variables.
}
\usage{
data(FC_raw)
}
\description{
This data set contains synthetic raw data of 1000 participants on a ranking
task on 15 items.
}
\keyword{datasets}
